Allow admins to delete quotes

+++ quotes/index.js
@@ -22,8 +22,8 @@ app.post("/api/add", checkUser, async (req, res) => {
 
 app.get("/api/list", checkUser, async (req, res) => {
     const quotes = await db.query(
-        "SELECT q.id, a.name, a.middlename, a.surname, q.quote, c.name AS class, (q.user_id = ?) AS owner FROM quotes AS q INNER JOIN users AS a ON author_id = a.id INNER JOIN class AS c ON a.class_id = c.id ORDER BY a.name",
-        [req.session.uid],
+        "SELECT q.id, a.name, a.middlename, a.surname, q.quote, c.name AS class, (q.user_id = ? OR ?) AS owner FROM quotes AS q INNER JOIN users AS a ON author_id = a.id INNER JOIN class AS c ON a.class_id = c.id ORDER BY a.name",
+        [req.session.uid, req.session.isAdmin],
     );
     res.json(quotes);
 });
@@ -31,7 +31,7 @@ app.get("/api/list", checkUser, async (req, res) => {
 app.delete("/api/delete/:id", checkUser, async (req, res) => {
     if (!req.params.id) return res.send("error");
     try {
-        await db.query("DELETE FROM quotes WHERE id = ? AND user_id = ?", [req.params.id, req.session.uid]);
+        await db.query("DELETE FROM quotes WHERE id = ? AND (user_id = ? OR ?)", [req.params.id, req.session.uid, req.session.isAdmin]);
         res.send("ok");
     } catch (e) {
         console.error(e);
